[CSBindings] Don't record variable adjacency for `member chain base` constraint
This constraint is used to enforce ordering of inference and provide
a starting point for inference of base based on associated generic
constraints. It's an `Equals` constraint with additional semantics
that only apply to unresolved members, which means that there are
no (other) adjacent variables in this case.

Resolves: rdar://78425221
(cherry picked from commit b88114248674aa90076169f4794000217bade579)