[ownership] Teach the load_borrow verifier that value_metatype of an address isn't a write.
Just a case where the verifier was too sensitive. The only places that we use
this in terms of optimization would just not optimize in this case, so no
miscompiles are possible. That being said, we should ensure that SILGen doesn't
hit a verifier error after it emits code.

rdar://78698170 [SR-14680]