[5.5] [Async Refactoring] Better handle non-refutable patterns (#37848)
* Explicitly handle error case in fallback logic

I don't believe this case currently can come up,
but leave it explicitly unhandled for now so we
don't perform an erroneous transform if it ever
comes up in the future.

* Add a couple of getSemanticsProviding calls

* Don't handle refutable patterns in async transform

These aren't currently supported.

* Don't handle unclassified switch cases for now

Previously we would silently add these to the
success block. For now, let's leave them
unhandled.

* Allow SemaAnnotator to handle patterns

Add the necessary walking hooks, and fix
ReferenceCollector to use it.

* Adapt ScopedDeclCollector to track number of references

Use a DenseMap to track the number of references to
a given decl in a scope.

* [AST] Add Pattern::hasAnyMutableBindings

* Generalise addTupleOf

Allow it to iterate over an arbitrary container
type.

* Better handle non-refutable patterns in async transform

Keep track of patterns that bind multiple vars and
print them out when converting an async call. If
the parameter being bound isn't referenced elsewhere,
we'll print the pattern inline as e.g:

```
let ((x, y), z) = await foo()
```

Otherwise, if the parameter is referenced elsewhere
in the block we'll print the pattern out of line,
such as:

```
let (res, z) = await foo()
let (x, y) = res
```

In addition, make sure to print var bindings out
of line if there's also a let binding, e.g:

```
let x = await foo()
var y = x
```

This ensures any mutations to y doesn't affect x.
If there's only a single var binding, we'll print
it inline.

rdar://77802560

* Handle parent vars in async transform

If we're in a case stmt body, any DeclRefExprs
to vars bound by a pattern will actually be to an
implicit var decl that's declared for the body. We
therefore need to walk to its "parent" to get to
the var referenced by the pattern, which will have
the correct entries in the naming and placeholder
maps.

* Strip an unused DiagnosticEngine param

* Add a release/5.5 workaround for a double AST walk

Avoid walking twice for capture lists. On main,
this is fixed in the AST walker.