Fix spurious weak lifetime diagnostic.
Borrowed values can still escape and be copied:

  %alloc = allocation
  %borrow = begin_borrow %alloc
  apply(%borrow) // escape
  end_borrow %borrow
  weak_store %alloc

Don't warn in this case.

While we're at it, handle all kinds of forwarding operations
consistently. Otherwise, it's far too easy for the anlysis to be
defeated, making it unlikely to catch bugs that cause spurious
warnings.

Fixes rdar://79146338 (Xcode warns that "referenced object is
deallocated here" but that object was passed into a method that causes
strong retention)