GSB: Don't accidentally create a GSB from a requirement signature in an assert
When a protocol's requirement signature is being computed by
GenericSignatureBuilder::computeGenericSignature(), we call
checkGenericSignature(), which contains various assertions
that call GenericSignature::isCanonicalTypeInContext().

One of these assertions forgot to pass the GSB instance down
as the 'builder' parameter.

The end result is that we would create a new GSB from the
protocol's requirement signature, which is generally not
well-formed since it does not have a conformance requirement
on 'Self'.

It seems this was harmless, other than the wasted CPU cycles,
but it was caught by some RequirementMachine assertions.