[Docs] Added OldMangling.rst (a copy of part of the old ABI.rst)
We still have code that generates the old mangling (in OldRemangler.cpp), so we
could really do with keeping the documentation for the old mangling scheme
around, at least for now.  To that end, I've fetched ABI.rst as at 76820edda97,
removed things not relating to mangling and moved it to ABI/OldMangling.rst.