Local functions can capture "self" as isolated.
This is a short-term fix to address a regression in actor-isolation
checking. A longer-term fix involves changing the way in which we
model actor isolation so that local functions can be stated to be
isolated to a particular capture, much like closures.

This subsumes my short-sighted fix for "defer".

Fixes rdar://79564012.