GSB: Remove a couple of well-formedness assertions
The underlying issue here won't be fixed in 5.4 and 5.5, so remove the
asserts to avoid allowing code in no-asserts builds that is rejected in
asserts builds.

Any code that builds with this assert commented out might fall victim
to an ABI break once the bug with same-type requirement minimization
is actually fixed.

"Fixes" https://bugs.swift.org/browse/SR-14776.