[SIL] Fix incorrect handling of 'forwarding' when parsing SIL
The `forwarding` directive should be optional for SIL instructions.
However, the current SIL parser put it as a requirement when there is a
comma follows after the main instruction components, which conflicts
with debug info directives like `loc` or `scope`. This patch teaches the
parser to treat `forwarding` as an optional directive.

Resolves SR-14814.